requirements.txt
streamlit
pillow
pytesseract
PyMuPDF
google-generativeai
python-dotenv

